% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mu_sigma.R
\name{calculate_mu_sigma}
\alias{calculate_mu_sigma}
\title{Calculate mu and sigma}
\usage{
calculate_mu_sigma(
  est_method = parent.frame()$est_method,
  losA = parent.frame()$losA,
  sd_los = parent.frame()$sd_los
)
}
\arguments{
\item{est_method}{Integer - indicates estimation method to use, either 1 or 2}

\item{losA}{Dataframe - imported from excel, contains length of stay
parameters}

\item{sd_los}{Float - standard deviation for length of stay distribution}
}
\value{
losA dataframe with new mu, sigma and los_params columns.
}
\description{
Calculates mu and sigma, and stores both within los_params, within the
length of stay dataframe (losA).
}
\examples{
\dontrun{
losA <- calculate_mu_sigma(est_method = 1, losA = losA, sd_los = 3)
}
}
