% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_df.R
\name{create_output_df}
\alias{create_output_df}
\title{Create output dataframe}
\usage{
create_output_df(nrow)
}
\arguments{
\item{nrow}{Integer - number of rows}
}
\value{
Dataframe
}
\description{
Create blank dataframe for output for visit-based simulation after warm-up
}
\examples{
create_output_df(nrow = 3)
}
