% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patient_df.R
\name{create_patient_df}
\alias{create_patient_df}
\title{Create patient dataframe}
\usage{
create_patient_df(nrow)
}
\arguments{
\item{nrow}{integer - number of rows}
}
\value{
dataframe
}
\description{
Create blank dataframe to populate for patients in visit-based simulation
}
\examples{
create_patient_df(nrow = 3)
}
