% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scenarios.R
\name{create_scenarios}
\alias{create_scenarios}
\title{Create scenarios}
\usage{
create_scenarios(type, arrivals_all, capacity, losA, init_conds)
}
\arguments{
\item{type}{string "scenarios" or NULL}

\item{arrivals_all}{dataframe}

\item{capacity}{dataframe}

\item{losA}{dataframe}

\item{init_conds}{dataframe}
}
\value{
scenarios dataframe
}
\description{
Combine arrivals_all, capacity, losA and init_conds
to create scenarios and arr_scenarios
}
