% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_df.R
\name{create_summary_df}
\alias{create_summary_df}
\title{Create summary dataframe}
\usage{
create_summary_df(nrow)
}
\arguments{
\item{nrow}{Integer - number of rows}
}
\value{
Dataframe
}
\description{
Create blank summary dataframe for visit-based simulation
}
\examples{
create_summary_df(nrow = 3)
}
