% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wait_df.R
\name{create_wait_df}
\alias{create_wait_df}
\title{Create wait time dataframe}
\usage{
create_wait_df()
}
\value{
dataframe
}
\description{
Create blank dataframe to store wait times for visit-based simulation
}
\examples{
create_wait_df()
}
