% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_end_slots.R
\name{dis_end_slots}
\alias{dis_end_slots}
\title{Sample end number of visits}
\usage{
dis_end_slots(
  z = parent.frame()$z,
  end_sr = parent.frame()$end_sr[z],
  sd_esr = parent.frame()$sd_esr[z]
)
}
\arguments{
\item{z}{string - scenario}

\item{end_sr}{float - mean for ESR distribution}

\item{sd_esr}{float - standard deviation for ESR distribution}
}
\value{
x integer - ESR
}
\description{
Sample from normal distribution to find the number of visits required on
patient's final day in pathway.
}
\details{
End service rate (ESR) or final visit rate (FVR) have a normal
distribution. Based on the provided mean and SD for that distribution,
samples to get ESR/FVR. It cannot be (a) less than 0, or (b) more than
mean+SD*3.
}
\examples{
dis_end_slots(z = "P1_B_BCap_Blos_BArr", end_sr = 1, sd_esr = 0.5)
}
