% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_los.R
\name{dis_los}
\alias{dis_los}
\title{Sample from length of stay (LOS) distribution}
\usage{
dis_los(
  z = parent.frame()$z,
  los_dist = parent.frame()$srv_dist_visit[[z]],
  los_lnorm_mean_sd = parent.frame()$srv_params_visit[[z]],
  los_norm_mean = parent.frame()$mean_los_visit[[z]],
  los_norm_sd = parent.frame()$sd_los_visit[[z]]
)
}
\arguments{
\item{z}{string - scenario}

\item{los_dist}{the distribution type (e.g. "lnorm") - default is from srv_dist_visit}

\item{los_lnorm_mean_sd}{the mean and SD for the lnorm distribution - default is from srv_params_visit}

\item{los_norm_mean}{the mean for the other distribution type - default is from mean_los_visit}

\item{los_norm_sd}{the SD for the other distribution type - default is from sd_los_visit}
}
\value{
as.integer(x) - the length of stay as an integer
}
\description{
Calculates maximum LOS from mean and SD of LOS distribution.
If LOS is lnorm, sample from rlnorm distribution until you get a value
that is >=0 and <= the max LOS. Uses do.call() as we need to input
mean and SD for rlnorm() from a list
}
\examples{
dis_los(z = "P1_B_BCap_Blos_BArr",
        los_dist = "lnorm",
        los_lnorm_mean_sd = c(2.197225, 1.180702),
        los_norm_mean = 18.07,
        los_norm_sd = 3)
}
