% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_los2.R
\name{dis_los2}
\alias{dis_los2}
\title{Shorten length of stay}
\usage{
dis_los2(long_los)
}
\arguments{
\item{long_los}{float - length of stay (output by dis_los())}
}
\value{
Integer - shortened length of stay
}
\description{
Reduces length of stay for patients already in service
}
\details{
Samples from uniform distribution between 1 and the length of stay
(produced by dis_los()). Rounds to nearest integer. Assumption: All patients
stay at least 1 day.
}
\examples{
dis_los2(18.34)
}
