% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import file (test function)}
\usage{
import_file(filename, sheet)
}
\arguments{
\item{filename}{string}

\item{sheet}{string}
}
\value{
file dataframe
}
\description{
Import file (test function)
}
