% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{load_setup}
\alias{load_setup}
\title{Load setup}
\usage{
load_setup(input_filename)
}
\arguments{
\item{input_filename}{A string (name of file with model inputs)}
}
\value{
Runs setup.R script from inst/scripts
}
\description{
Load setup
}
