% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_visit_sim.R
\name{run_visit_sim}
\alias{run_visit_sim}
\title{Run visit-based simulation}
\usage{
run_visit_sim(
  pathway_vector_visit = parent.frame()$pathway_vector_visit,
  nruns = parent.frame()$nruns,
  temp_seed = parent.frame()$temp_seed,
  sim_length = parent.frame()$sim_length,
  warmup = parent.frame()$warmup,
  n_slots = parent.frame()$n_slots,
  init_occ_visit = parent.frame()$init_occ_visit,
  init_niq_visit = parent.frame()$init_niq_visit,
  arr_rates_visit = parent.frame()$arr_rates_visit,
  isr = parent.frame()$isr,
  end_sr = parent.frame()$end_sr,
  mean_los_visit = parent.frame()$mean_los_visit,
  costs_visit = parent.frame()$costs_visit,
  srv_dist_visit = parent.frame()$srv_dist_visit,
  srv_params_visit = parent.frame()$srv_params_visit,
  sd_los_visit = parent.frame()$sd_los_visit,
  sd_isr = parent.frame()$sd_isr,
  sd_esr = parent.frame()$sd_esr
)
}
\arguments{
\item{pathway_vector_visit}{tbc}

\item{nruns}{tbc}

\item{temp_seed}{tbc}

\item{sim_length}{tbc}

\item{warmup}{tbc}

\item{n_slots}{tbc}

\item{init_occ_visit}{tbc}

\item{init_niq_visit}{tbc}

\item{arr_rates_visit}{tbc}

\item{isr}{tbc}

\item{end_sr}{tbc}

\item{mean_los_visit}{tbc}

\item{costs_visit}{tbc}

\item{srv_dist_visit}{tbc}

\item{srv_params_visit}{tbc}

\item{sd_los_visit}{tbc}

\item{sd_isr}{tbc}

\item{sd_esr}{tbc}
}
\value{
List containing visits_based_output and visits_based_output_q
}
\description{
Run visit-based simulation
}
