% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_visit.R
\name{save_visit}
\alias{save_visit}
\title{Save output from visit-based simulation}
\usage{
save_visit(
  visit_filename,
  arr_rates_visit = parent.frame()$arr_rates_visit,
  outcomes = parent.frame()$outcomes,
  visits_based_output = parent.frame()$visits_based_output
)
}
\arguments{
\item{visit_filename}{string - path to save output csv to}

\item{arr_rates_visit}{tbc}

\item{outcomes}{tbc}

\item{visits_based_output}{tbc}
}
\value{
saves csv to specified location
}
\description{
Extract results for each date for number in queue, occupancy, wait and costs and then save to csv
}
\details{
Extract results from visit-based simulation output for NIQ, OCC, wait and
costs, then append together, along with dates. Save this to a csv using the
output filename specified.
}
