% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_visit_stoch.R
\name{save_visit_stoch}
\alias{save_visit_stoch}
\title{Save stochastic output from visit-based simulation}
\usage{
save_visit_stoch(
  visit_stoch_filename,
  visits_based_output_q = parent.frame()$visits_based_output_q
)
}
\arguments{
\item{visit_stoch_filename}{String - path to save output csv to}

\item{visits_based_output}{Dataframe with stochastic results from
visit-based simulation}
}
\value{
Saves csv to specified location
}
\description{
Correct date formatting, and save the quantiles to csv for optional
stochastic report, saving in long format for plotting
}
