% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_all.R
\name{setup_all}
\alias{setup_all}
\title{Set up for simulation}
\usage{
setup_all(model_type)
}
\arguments{
\item{model_type}{A string (either "visit" or "bed")}
}
\value{
list(return_new_names, return_objects). Return_new_names is a list
of strings, which are the names of objects in return_objects. Return_objects
contains:
\itemize{
\item scenarioes - dataframe (need to add description for each of these)
\item costs
\item init_occ
\item init_niq
\item srv_dist
\item cap
\item loss
\item srv_params
\item mean_los
\item sd_los
\item arr_rates
\item pathway_vector
}
}
\description{
Set up for simulation
}
\examples{
setup_visit <- setup_all("visit")
setup_visit <- setup_all("bed")
}
