% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_patient.R
\name{add_patient}
\alias{add_patient}
\title{Add patient}
\usage{
add_patient(
  in_system,
  id = parent.frame()$id,
  npat = parent.frame()$npat,
  req_visits = parent.frame()$req_visits,
  patients = parent.frame()$patients,
  resources = parent.frame()$resources
)
}
\arguments{
\item{in_system}{boolean (TRUE or FALSE) - is the patient already in the
system? If so, will produce adjusted length of stay and visit rates.}

\item{id}{integer}

\item{npat}{integer}

\item{req_visits}{tbc}

\item{patients}{dataframe}

\item{resources}{dataframe}
}
\value{
List with updated id, npat, req_visits, patients, resources objects
}
\description{
Add patient to visit-based simulation pathway
}
\details{
When run add_patient(), it will increment the ID and n_pat, find length of
stay, initial visit rate and end visit rate. It will save these results in a
dataframe, and then check availability in service and modify resources
accordingly.
}
