% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mu_sigma.R
\name{calculate_mu_sigma}
\alias{calculate_mu_sigma}
\title{Calculate mu and sigma}
\usage{
calculate_mu_sigma(
  est_method = 1,
  losA = parent.frame()$losA,
  sd_los = parent.frame()$sd_los
)
}
\arguments{
\item{est_method}{integer (should be set to 1 or 2). Default 1.}

\item{losA}{dataframe (with length of stay parameters, imported from excel)}

\item{sd_los}{number (SD for length of stay). Default 3.}
}
\value{
losA
}
\description{
Calculates mu and sigma, and stores both within los_params
}
