% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mu_sigma.R
\name{calculate_mu_sigma}
\alias{calculate_mu_sigma}
\title{Calculate mu and sigma}
\usage{
calculate_mu_sigma(
  est_method = parent.frame()$est_method,
  losA = parent.frame()$losA,
  sd_los = parent.frame()$sd_los
)
}
\arguments{
\item{est_method}{Integer. Indicates estimation method to use, either 1 or 2.}

\item{losA}{Dataframe.}

\item{sd_los}{Float. Standard deviation for length of stay distribution.}
}
\value{
losA with new mu, sigma and los_params columns.
}
\description{
Calculates mu and sigma, and stores both within los_params, within the
length of stay dataframe (losA).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
losA <- calculate_mu_sigma(est_method = 1, losA = losA, sd_los = 3)
\dontshow{\}) # examplesIf}
}
