% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import simulation excel file}
\usage{
import_file(filename)
}
\arguments{
\item{filename}{String.}
}
\value{
List of length 2. Contains list of dataframe names and list of
dataframes.
}
\description{
Imports each sheet from the excel file that contain parameters for the
simulation. File should contain sheets "arrivals", "initial conditions",
"capacity", "los" and "costs". Returns these as dataframes, along with
list of dataframe names.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
import_file("ipacs_model_parameters.xlsx")
\dontshow{\}) # examplesIf}
}
