% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Imports the sheets from the excel file with the model parameters
File should contain sheets "arrivals", "initial conditions", "capacity",
"los" and "costs"}
\usage{
import_file(filename)
}
\arguments{
\item{filename}{string}
}
\value{
list(df_names, df_list)
}
\description{
Imports the sheets from the excel file with the model parameters
File should contain sheets "arrivals", "initial conditions", "capacity",
"los" and "costs"
}
