% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_visit.R
\name{save_visit}
\alias{save_visit}
\title{Save standard output from visit-based simulation}
\usage{
save_visit(
  visit_filename,
  arr_rates_visit = parent.frame()$arr_rates_visit,
  visits_based_output = parent.frame()$visits_based_output
)
}
\arguments{
\item{visit_filename}{string - path to save output csv to}

\item{arr_rates_visit}{Dataframe - first column is date, subsequent columns
are each scenario, with elements of dataframe being the arrival rate for
each scenario and date, e.g. data.frame(date = as.Date(c("2023-01-01",
"2023-01-02")), P1_B_BCap_Blos_BArr = c(3.790697, 3.047619))}

\item{visits_based_output}{Dataframe with results from visit-based simulation}
}
\value{
Saves csv to specified location
}
\description{
Extract results from visit-based simulation output for NIQ, OCC, wait and
costs, then append together, along with dates. Save this to a csv using the
output filename specified.
}
